add  back inline buttons

+++ schedule.py
@@ -8,24 +8,33 @@ from proj.tasks import push_message_at_day
 
 
 async def do_every_seconds(second: int, chat_id):
-    schedule.every(second).seconds.do(push_message,
-                                      chat=chat_id).tag(str(chat_id))
-    await FrequencyPosting.filter(chat_id=chat_id).delete()
-    await FrequencyPosting.create(chat_id=chat_id, count_time=second, type_time='секунд', type='by_period')
+    if second > 0:
+        schedule.every(second).seconds.do(push_message,
+                                        chat=chat_id).tag(str(chat_id))
+        await FrequencyPosting.filter(chat_id=chat_id).delete()
+        await FrequencyPosting.create(chat_id=chat_id, count_time=second, type_time='секунд', type='by_period')
+    else:
+        raise ValueError
 
 
 async def do_every_minutes(minute, chat_id):
-    schedule.every(minute).minutes.do(push_message,
-                                      chat=chat_id).tag(str(chat_id))
-    await FrequencyPosting.filter(chat_id=chat_id).delete()
-    await FrequencyPosting.create(chat_id=chat_id, count_time=minute, type_time='минут', type='by_period')
+    if minute > 0:
+        schedule.every(minute).minutes.do(push_message,
+                                        chat=chat_id).tag(str(chat_id))
+        await FrequencyPosting.filter(chat_id=chat_id).delete()
+        await FrequencyPosting.create(chat_id=chat_id, count_time=minute, type_time='минут', type='by_period')
+    else:
+        raise ValueError
 
 
 async def do_every_hours(hour, chat_id):
-    schedule.every(hour).hours.do(push_message,
-                                  chat=chat_id).tag(str(chat_id))
-    await FrequencyPosting.filter(chat_id=chat_id).delete()
-    await FrequencyPosting.create(chat_id=chat_id, count_time=hour, type_time='часов', type='by_period')
+    if hour > 0:
+        schedule.every(hour).hours.do(push_message,
+                                    chat=chat_id).tag(str(chat_id))
+        await FrequencyPosting.filter(chat_id=chat_id).delete()
+        await FrequencyPosting.create(chat_id=chat_id, count_time=hour, type_time='часов', type='by_period')
+    else:
+        raise ValueError
 
 
 async def set_frequency_posting(chat_id: int, type_frequency: str, frequency: str):
@@ -41,9 +50,11 @@ async def set_frequency_posting(chat_id: int, type_frequency: str, frequency: st
             await frequency_dict[frequency_list[1]](int(frequency_list[0]), chat_id)
         except KeyError:
             return 'Неправильный формат число:тип (без пробелов)'
+        except ValueError:
+            return "Введённое число должно быть больше 0!"
     if type_frequency == 'by_day':
         if 7 < int(frequency) < 1:
-            return "Введенное число должно быть в диапазоне от 1 до 7"
+            return "Введенное число должно быть в диапазоне от 1 до 7!"
         frequency_dict = {
             '1': schedule.every().monday.do(push_message,
                                             chat=chat_id).tag(str(chat_id)),
@@ -70,8 +81,11 @@ async def set_frequency_posting(chat_id: int, type_frequency: str, frequency: st
             push_message_at_day.apply_async((chat_id, await get_first_message(chat_id)),
                                             eta=datetime.today().replace(day=int(frequency)))
             message = await MessageQueue.filter(chat_id=chat_id).first()
-            await MessageQueue.filter(id=message.id).delete()
-            await FrequencyPosting.filter(chat_id=chat_id).delete()
+            await message.delete()
+            try:
+                await FrequencyPosting.filter(chat_id=chat_id).delete()
+            except:
+                pass
             await FrequencyPosting.create(chat_id=chat_id, type='by_date', day_of_month=int(frequency))
         except ValueError:
             return "День должен быть в диапазоне от 1 до 31"
